Change from long to int.

+++ sdk/eventhubs/azure-messaging-eventhubs/src/main/java/com/azure/messaging/eventhubs/EventHubMessageSerializer.java
@@ -187,16 +187,16 @@ class EventHubMessageSerializer implements MessageSerializer {
         }
 
         final Map<Symbol, Object> deliveryAnnotations = annotations.getValue();
-        final Long lastSequenceNumber = getValue(deliveryAnnotations, 
+        final Long lastSequenceNumber = getValue(deliveryAnnotations,
             MANAGEMENT_RESULT_LAST_ENQUEUED_SEQUENCE_NUMBER, Long.class);
-        final String lastEnqueuedOffset = getValue(deliveryAnnotations, 
+        final String lastEnqueuedOffset = getValue(deliveryAnnotations,
             MANAGEMENT_RESULT_LAST_ENQUEUED_OFFSET, String.class);
         final Instant lastEnqueuedTime = getValue(deliveryAnnotations,
             MANAGEMENT_RESULT_LAST_ENQUEUED_TIME_UTC, Date.class).toInstant();
-        final Instant retrievalTime = getValue(deliveryAnnotations, 
+        final Instant retrievalTime = getValue(deliveryAnnotations,
             MANAGEMENT_RESULT_RUNTIME_INFO_RETRIEVAL_TIME_UTC, Date.class).toInstant();
-        final Long lastEnqueuedReplicationSegment = getValue(deliveryAnnotations,
-            MANAGEMENT_RESULT_LAST_ENQUEUED_SEQUENCE_NUMBER_EPOCH, Long.class, false);
+        final Integer lastEnqueuedReplicationSegment = getValue(deliveryAnnotations,
+            MANAGEMENT_RESULT_LAST_ENQUEUED_SEQUENCE_NUMBER_EPOCH, Integer.class, false);
 
         return new LastEnqueuedEventProperties(lastSequenceNumber, Long.valueOf(lastEnqueuedOffset), lastEnqueuedTime,
             retrievalTime, lastEnqueuedReplicationSegment);
@@ -219,7 +219,7 @@ class EventHubMessageSerializer implements MessageSerializer {
                 "enqueuedTime: %s should always be in map.", SEQUENCE_NUMBER_ANNOTATION_NAME.getValue())));
         }
 
-        final Instant enqueuedTime = MessageUtils.getEnqueuedTime(messageAnnotations, 
+        final Instant enqueuedTime = MessageUtils.getEnqueuedTime(messageAnnotations,
             ENQUEUED_TIME_UTC_ANNOTATION_NAME.getValue());
 
         final String partitionKey = (String) messageAnnotations.get(PARTITION_KEY_ANNOTATION_NAME.getValue());
@@ -227,13 +227,13 @@ class EventHubMessageSerializer implements MessageSerializer {
         final long sequenceNumber = getAsLong(messageAnnotations, SEQUENCE_NUMBER_ANNOTATION_NAME.getValue());
 
         // It is an optional value. Possible that there is no replication segment.
-        final Long replicationSegment = (Long) messageAnnotations.get(REPLICATION_SEGMENT_ANNOTATION_NAME.getValue());
+        final Integer replicationSegment = (Integer) messageAnnotations.get(REPLICATION_SEGMENT_ANNOTATION_NAME.getValue());
 
         // Put the properly converted time back into the dictionary.
         messageAnnotations.put(OFFSET_ANNOTATION_NAME.getValue(), offset);
         messageAnnotations.put(ENQUEUED_TIME_UTC_ANNOTATION_NAME.getValue(), enqueuedTime);
         messageAnnotations.put(SEQUENCE_NUMBER_ANNOTATION_NAME.getValue(), sequenceNumber);
-        
+
         messageAnnotations.put(REPLICATION_SEGMENT_ANNOTATION_NAME.getValue(), replicationSegment);
 
         LOGGER.atInfo()
@@ -262,8 +262,8 @@ class EventHubMessageSerializer implements MessageSerializer {
             getValue(amqpBody, ManagementChannel.MANAGEMENT_RESULT_LAST_ENQUEUED_OFFSET, String.class),
             getDate(amqpBody, ManagementChannel.MANAGEMENT_RESULT_LAST_ENQUEUED_TIME_UTC),
             getValue(amqpBody, ManagementChannel.MANAGEMENT_RESULT_PARTITION_IS_EMPTY, Boolean.class),
-            getValue(amqpBody, ManagementChannel.MANAGEMENT_RESULT_BEGINNING_SEQUENCE_NUMBER_EPOCH, Long.class, false),
-            getValue(amqpBody, MANAGEMENT_RESULT_LAST_ENQUEUED_SEQUENCE_NUMBER_EPOCH, Long.class, false));
+            getValue(amqpBody, ManagementChannel.MANAGEMENT_RESULT_BEGINNING_SEQUENCE_NUMBER_EPOCH, Integer.class, false),
+            getValue(amqpBody, MANAGEMENT_RESULT_LAST_ENQUEUED_SEQUENCE_NUMBER_EPOCH, Integer.class, false));
     }
 
     /**
